requirements.txt
pyyaml>=6.0
arxiv>=2.1.0
arxiv-to-prompt>=0.11.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
